requirements.txt
amqp==5.1.0
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.11
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangorestframework-xml==2.0.0
drf-yasg==1.20.0
executing==0.8.2
Faker==11.3.0
flake8==4.0.1
flower==1.0.0
humanize==4.0.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
ipython==7.31.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prometheus-client==0.14.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
text-unidecode==1.3
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
